Initial project tree setup, tests passed. - Files renamed/created - Some files moved
