+++ Gemfile
@@ -2,6 +2,11 @@
 
 source 'https://rubygems.org'
 
+# Web Application
+gem 'puma', '~> 5'
+gem 'roda', '~> 3'
+gem 'slim', '~> 4'
+
 # Networking
 gem 'http', '~> 5'
 
@@ -12,8 +17,15 @@ gem 'simplecov', '~> 0'
 gem 'vcr', '~> 6'
 gem 'webmock', '~> 3'
 
+# Validation
+gem 'dry-struct', '~> 1'
+gem 'dry-types', '~> 1'
+
 # Utility Tools
 gem 'rake'
+
+# Debugging
+gem 'pry'
 
 # Code Quality
 gem 'flog'
