+++ Rakefile
@@ -10,7 +10,7 @@ end
 
 desc 'Run tests'
 task :spec do
-  sh 'ruby spec/amadeus_api_spec.rb'
+  sh 'ruby spec/gateway_amadeus_api_spec.rb'
 end
 
 namespace :vcr do
@@ -40,4 +40,9 @@ namespace :quality do
   task :flog do
     sh "flog #{CODE}"
   end
+end
+
+desc 'Starts web app'
+task :run do
+  sh 'bundle exec puma'
 end
