+++ config.ru
@@ -1,0 +1,4 @@
+require_relative '../require_app'
+require_app
+
+run ComfyWings::App.freeze.app
