+++ require_app.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# Require all ruby files in the following folders
+def require_app
+  Dir.glob('./{config,app}/**/*.rb').each do |file|
+    require file
+  end
+end
