+++ spec/fixtures/airport_info.rb
