+++ spec/fixtures/flight_info.rb
