+++ spec/gateway_amadeus_api_spec.rb
@@ -7,8 +7,8 @@ describe 'Tests AMADEUS API library' do
     c.cassette_library_dir = CASSETTES_FOLDER
     c.hook_into :webmock
 
-    c.filter_sensitive_data('<AMADEUS_TOKEN>') { AMADEUS_TOKEN }
-    c.filter_sensitive_data('<AMADEUS_TOKEN_ESC>') { CGI.escape(AMADEUS_TOKEN) }
+    c.filter_sensitive_data('<AMADEUS_KEY>') { AMADEUS_KEY }
+    c.filter_sensitive_data('<AMADEUS_KEY_ESC>') { CGI.escape(AMADEUS_KEY) }
 
     c.filter_sensitive_data('<AMADEUS_SECRET>') { AMADEUS_SECRET }
     c.filter_sensitive_data('<AMADEUS_SECRET_ESC>') { CGI.escape(AMADEUS_SECRET) }
@@ -27,34 +27,34 @@ describe 'Tests AMADEUS API library' do
 
   describe 'Flight information' do
     it 'HAPPY: should provide correct flight attributes' do
-      matched_flights = Amadeus::AmadeusApi.new(AMADEUS_TOKEN, AMADEUS_SECRET)
+      matched_flights = ComfyWings::Amadeus::Api.new(AMADEUS_KEY, AMADEUS_SECRET)
                                            .flight('TPE', 'MAD', '2022-11-01', '2022-11-05')
       _(matched_flights.size).must_equal CORRECT['flight_num']
     end
 
     it 'SAD: should raise exception on incorrect flight information' do
       _(proc do
-          Amadeus::AmadeusApi.new('BAD_TOKEN', 'BAD_SECRET')
+          ComfyWings::Amadeus::Api.new('BAD_TOKEN', 'BAD_SECRET')
                              .flight('AIRPORT_NOT_EXIST', 'AIRPORT_NOT_EXIST', '2022-11-01', '2022-11-05')
-        end).must_raise Amadeus::AmadeusApi::Response::Unauthorized
+        end).must_raise ComfyWings::Amadeus::Api::Response::Unauthorized
     end
 
     it 'SAD: should raise exception when unauthorized' do
       _(proc do
-        Amadeus::AmadeusApi.new('BAD_TOKEN', 'BAD_SECRET')
+        ComfyWings::Amadeus::Api.new('BAD_TOKEN', 'BAD_SECRET')
                            .flight('TPE', 'MAD', '2022-11-01', '2022-11-05')
-      end).must_raise Amadeus::AmadeusApi::Response::Unauthorized
+      end).must_raise ComfyWings::Amadeus::Api::Response::Unauthorized
     end
   end
 
   describe 'Flight information' do
     before do
-      @matched_flights = Amadeus::AmadeusApi.new(AMADEUS_TOKEN, AMADEUS_SECRET)
+      @matched_flights = ComfyWings::Amadeus::Api.new(AMADEUS_KEY, AMADEUS_SECRET)
                                             .flight('TPE', 'MAD', '2022-11-01', '2022-11-05')
     end
 
     it 'HAPPY: should recognize flight' do
-      _(@matched_flights.flights[0]).must_be_kind_of Amadeus::Flight
+      _(@matched_flights.flights[0]).must_be_kind_of ComfyWings::Amadeus::Flight
     end
 
     it 'HAPPY: should identify flight' do
