+++ spec/spec_helper.rb
@@ -10,13 +10,13 @@ require 'minitest/rg'
 require 'vcr'
 require 'webmock'
 
-require_relative '../lib/amadeus_api'
-require_relative '../lib/flight_info'
+require_relative '../app/models/gateways/amadeus_api'
+require_relative '../spec/fixtures/flight_info'
 
 # TODO: add/move required libraries, constants, and files required for tests
 #       when you write or create a new test
 CONFIG = YAML.safe_load(File.read('config/secrets.yml'))
-AMADEUS_TOKEN = CONFIG['AMADEUS_KEY']
+AMADEUS_KEY = CONFIG['AMADEUS_KEY']
 AMADEUS_SECRET = CONFIG['AMADEUS_SECRET']
 CORRECT = YAML.safe_load(File.read('spec/fixtures/flight_results.yml'))
